Added Cannare RCAP RCA plug images
The Canare RCAP plugs are a different breed, and have more in common
with the crimp BNC plugs widely used in the broadcast industry than
they do with solder-type RCA plugs. In the Canare RCAP, the coaxial
cable, still encased in its dielectric, enters the rear of the
connector through a perfectly round tunnel of nickel-plated brass;
because this tunnel is sized for the specific cable diameter (unlike
solder-type plugs which often are "one size fits all," the RCAP is made
in numerous different sizes) and is crimped to the coax shield, it
effectively replaces the stripped-away shield of the coax, continuing
the cable's 75 ohm impedance along toward the connector tip. At the end
of the cable dielectric, the center conductor is crimped to a
gold-plated pin which fits a socket centered in the plug body to make
contact with the RCA plug pin.

Where an ordinary solder-type RCA plug becomes asymmetrical, and
departs dramatically from the required dimensions to maintain a 75 ohm
impedance, well before the "business" end of the plug, the Canare RCA
plug maintains the coaxial geometry through its entire length, and
maintains a 75 ohm impedance as far as possible into the plug. While
it's inaccurate to call this a "true 75 ohm" RCA plug, it's accurate to
say that no other plug on the market does such a good job of minimizing
the effects of the impedance mismatch caused by the RCA plug's
unfortunate dimensions. The RCAP's reliability, durability, high pull
strength, and excellent spring-grip outer contact complete the package,
making it, in our opinion, far and away the best RCA plug available for
video use.